Make newlfs work by adding the MAXSYMLINKLEN and a byte padding to struct lfs.
sovs@diku.dk:
The pad2 field in struct lfs is a kludge. You
need it because the raw devices has to be written in DEV_BSIZE. The
lfs_maxsymlinklen is needed because the tells the UFS code that we
want to use 4.4BSD directory layout insted of FreeBSD 1.1.5.

Reviewed by: Justin Gibbs
Submitted by: sovs@diku.dk