Define MASTER_SITE_SUBDIR and PATCH_SITE_SUBDIR to default to "." so we won't have double-slashes.
Add support for new port variable MANUAL_PACKAGE_BUILD.  If this is
defined as well as the user variable PACKAGE_BUILDING, the port will
be ignored.  This is used to mark ports that can be built normally
except on a machine that has a lot of conflicting ports (i.e., our
package building machine).