Use KMODDEPS to document module dependencies on libiconv and libmchain.
KMODDEPS seems to be a no-op in HEAD, but is required in RELENG_4,
where MODULE_DEPEND seems to be a no-op.

Therefore, this change is harmless in -CURRENT, but will fix the
dependencies when merged to RELENG_4, where they are currently not
registered!

PR:		kern/33625
Submitted by:	Joe Marcus Clarke <marcus@marcuscom.com>