Make the ENV configuration variable a hash rather than an array.
Build LINT on -STABLE now that tinderbox.pl knows how.  Also try to build
LINT on powerpc and amd64 (this is a formality as they don't have NOTES
so nothing will be built)

Add two setups for release testing, with plenty of NO* to speed things up.

If the config key was not specified on the command line, try to guess it
from the hostname.